OBO error handlings and cleaned up test-bots\OBOAuthorizationBot
