+++ src/libraries/BotBuilder/Microsoft.Agents.BotBuilder/Properties/Resources.resx
@@ -162,11 +162,15 @@
     <comment>Raised when user sign in is already active.</comment>
   </data>
   <data name="OBONotExchangeableToken" xml:space="preserve">
-    <value>OBO TokenExchange for '{0}' cannot exchange an application that does not have an api:// audience.</value>
+    <value>OBO for '{0}' cannot exchange an application that does not have an api:// audience.</value>
     <comment>Raised when the Azure Bot OAuth Connection is not configured to return an exchangeable token.</comment>
   </data>
   <data name="OBONotSupported" xml:space="preserve">
-    <value>OBO Exchange not supported on '{0}'</value>
+    <value>OBO not supported on '{0}'</value>
     <comment>Raised when User Authorization is using a provider that does not support IOBOExchange</comment>
+  </data>
+  <data name="OBOExchangeFailed" xml:space="preserve">
+    <value>OBO exchange failed for connection '{0}' with scopes {1}.</value>
+    <comment>Raised when an OBO exchange was attempted.</comment>
   </data>
 </root>
