+++ src/samples/test-bots/OBOAuthorizationBot/OBOAuthorizationBot.csproj
