+++ src/samples/test-bots/OBOAuthorizationBot/Program.cs
@@ -1,14 +1,13 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using AuthenticationBot;
 using Microsoft.Agents.Hosting.AspNetCore;
 using Microsoft.Agents.Samples;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using System.Threading.Tasks;
+using OBOAuthorizationBot;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -23,10 +22,10 @@ builder.Logging.AddDebug();
 builder.Services.AddBotAspNetAuthentication(builder.Configuration);
 
 // Add AgentApplicationOptions from config.
-builder.AddAgentApplicationOptions(autoSignInSelector: (context, cancellationToken) => Task.FromResult(context.Activity.Text == "auto"));
+builder.AddAgentApplicationOptions();
 
 // Add the bot (which is transient)
-builder.AddBot<AuthBot>();
+builder.AddBot<OBOAuth>();
 
 
 var app = builder.Build();
