+++ src/samples/test-bots/OBOAuthorizationBot/appManifest/color.png
