+++ src/samples/test-bots/OBOAuthorizationBot/appManifest/m365copilot-manifest.json
