+++ src/samples/test-bots/OBOAuthorizationBot/appManifest/outline.png
