+++ src/samples/test-bots/OBOAuthorizationBot/appManifest/teams-manifest.json
