+++ src/samples/test-bots/OBOAuthorizationBot/appsettings.json
@@ -5,13 +5,8 @@
     "NormalizeMentions": false,
 
     "UserAuthorization": {
-      "Default": "graph",
+      "AutoSignIn":  false,
       "Handlers": {
-        "graph": {
-          "Settings": {
-            "AzureBotOAuthConnectionName": null
-          }
-        },
         "obo": {
           "Settings": {
             "AzureBotOAuthConnectionName": "obo",
